rename folder to bedrock-agent
